chore(sveltekit): refactor the rest of the common server-side code
