+++ packages/sveltekit/src/server/index.ts
@@ -123,10 +123,10 @@ export * from '@sentry/node';
 // -------------------------
 // SvelteKit SDK exports:
 export { init } from './sdk';
-export { handleErrorWithSentry } from './handleError';
-export { wrapLoadWithSentry, wrapServerLoadWithSentry } from './load';
+export { handleErrorWithSentry } from '../server-common/handleError';
+export { wrapLoadWithSentry, wrapServerLoadWithSentry } from '../server-common/load';
 export { sentryHandle } from './handle';
-export { wrapServerRouteWithSentry } from './serverRoute';
+export { wrapServerRouteWithSentry } from '../server-common/serverRoute';
 
 /**
  * Tracks the Svelte component's initialization and mounting operation as well as
