+++ packages/sveltekit/src/server/sdk.ts
@@ -3,7 +3,7 @@ import type { NodeClient, NodeOptions } from '@sentry/node';
 import { getDefaultIntegrations as getDefaultNodeIntegrations } from '@sentry/node';
 import { init as initNodeSdk } from '@sentry/node';
 
-import { rewriteFramesIntegration } from './rewriteFramesIntegration';
+import { rewriteFramesIntegration } from '../server-common/rewriteFramesIntegration';
 
 /**
  *
