+++ packages/sveltekit/src/worker/handle.ts
@@ -1,7 +1,9 @@
 import { CloudflareOptions, continueTrace, wrapRequestHandler } from '@sentry/cloudflare';
+import { getDefaultIntegrations as getDefaultCloudflareIntegrations } from '@sentry/cloudflare';
 import type { Handle } from '@sveltejs/kit';
 
 import { sentryHandleGeneric, SentryHandleOptions } from '../server-common/handle';
+import { rewriteFramesIntegration } from '../server-common/rewriteFramesIntegration';
 
 /**
  * A SvelteKit handle function that wraps the request for Sentry error and
@@ -29,15 +31,18 @@ export function sentryHandle(handlerOptions?: SentryHandleOptions): Handle {
 /** Initializes Sentry SvelteKit Cloudflare SDK
  *  This should be before the sentryHandle() call.
  * */
-export function initCloudflareSentryHandle(handlerOptions: CloudflareOptions): Handle {
-  const options = { ...handlerOptions };
+export function initCloudflareSentryHandle(options: CloudflareOptions): Handle {
+  const opts: CloudflareOptions = {
+    defaultIntegrations: [...getDefaultCloudflareIntegrations(options), rewriteFramesIntegration()],
+    ...options,
+  };
 
   const handleInitSentry: Handle = ({ event, resolve }) => {
     // if event.platform exists (should be there in a cloudflare worker), then do the cloudflare sentry init
     return event.platform
       ? wrapRequestHandler(
         {
-          options,
+          options: opts,
           request: event.request,
           // @ts-expect-error This will exist in Cloudflare
           context: event.platform.context,
